+++ src/Behat/Context/Domain/AdminContext.php
@@ -1,17 +1,11 @@
 <?php
 
-/*
- * This file is part of the Proximum Vimeet project.
- *
- * Copyright (C) Proximum
- *
- * @author Elao <contact@elao.com>
- */
-
 namespace Proximum\Vimeet\Behat\Context\Domain;
 
 use Behat\Behat\Context\Context;
 use Proximum\Vimeet\Behat\Context\Domain\Proxy\AdminContextProxyInterface;
+use Proximum\Vimeet\Domain\Model\Admin;
+use Proximum\Vimeet\Domain\Model\Event;
 
 class AdminContext implements Context
 {
@@ -31,10 +25,40 @@ class AdminContext implements Context
      *
      * @param string $email
      */
-    public function createSuperAdmin(string $email)
+    public function createSuperAdmin(string $email): void
     {
-        $user = $this->adminContextProxy->getAdminManager()->create($email, 'ROLE_SUPER_ADMIN');
+        $this->createAdmin($email, 'ROLE_SUPER_ADMIN');
+    }
 
-        $this->adminContextProxy->getStorage()->set('admin', $user);
+    /**
+     * @Given /^the admin "(?P<email>[^"]+)" with role "(?P<role>[^"]+)" is created$/
+     *
+     * @param string $email
+     * @param string $role
+     */
+    public function createAdmin(string $email, string $role): void
+    {
+        $admin = $this->adminContextProxy->getAdminManager()->create($email, $role);
+
+        $this->adminContextProxy->getStorage()->set('admin', $admin);
+    }
+
+    /**
+     * @Given /^this admin can access this event$/
+     */
+    public function assignEvent(): void
+    {
+        $admin = $this->adminContextProxy->getStorage()->get('admin');
+        $event = $this->adminContextProxy->getStorage()->get('event');
+
+        if (!$admin instanceof Admin) {
+            Throw new \InvalidArgumentException('Admin not found');
+        }
+
+        if (!$event instanceof Event) {
+            Throw new \InvalidArgumentException('Event not found');
+        }
+
+        $this->adminContextProxy->getAdminManager()->assignEvent($admin, $event);
     }
 }