+++ CHANGELOG.md
@@ -10,6 +10,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - MV-242 - Fix meeting / contact evaluations on event dashboard
 - MV-243 - Add media link in sheet pdf printing
 
+### Added
+- VIMEET-2149 - Change Admin password algorithm to argon2i and increase length
+
 ## [2.35.1] - 2020-06-30
 ### Fixed
 - Hotfix - Get / export many happening questions instead of one or null